12/12/19 cleaned up most comments, fixed object display, realtime working
